refactor: Split static file & modify private location
