+++ examples/http-trigger/s.yaml
@@ -4,7 +4,7 @@ access: aliyun-release  #  秘钥别名
 
 services:
   fc-deploy-test: #  服务名称
-    component: fc-deploy  # 组件名称
+    component: ${path(../..)}  # 组件名称
     props: #  组件的属性值
       region: cn-hangzhou
       service:
